+++ frontend/src/lib/components/graph/FlowGraph.svelte
@@ -12,7 +12,7 @@
 		type Loop,
 		type Branch,
 		type NestedNodes,
-		type GraphModuleStates
+		type GraphModuleState
 	} from '.'
 	import { defaultIfEmptyString, encodeState } from '$lib/utils'
 	import { createEventDispatcher, onMount, setContext } from 'svelte'
@@ -31,7 +31,7 @@
 	export let minHeight: number = 0
 	export let maxHeight: number | undefined = undefined
 	export let notSelectable = false
-	export let flowModuleStates: Record<string, GraphModuleStates> | undefined = undefined
+	export let flowModuleStates: Record<string, GraphModuleState> | undefined = undefined
 	export let rebuildOnChange: any = undefined
 
 	export let selectedId: Writable<string | undefined> = writable<string | undefined>(undefined)
@@ -151,9 +151,7 @@
 				Object.entries(flowModuleStates ?? [])
 					.filter(([k, v]) => k.startsWith('failure'))
 					.forEach(([k, v]) => {
-						nestedNodes.push(
-							createErrorHandler({ id: k } as FlowModule, v.states[v.job_id].parent_module)
-						)
+						nestedNodes.push(createErrorHandler({ id: k } as FlowModule, v.parent_module))
 					})
 			}
 			const flatNodes = flattenNestedNodes(nestedNodes)
@@ -320,7 +318,6 @@
 		branchable: boolean,
 		modules: FlowModule[]
 	): Node {
-		let fms = flowModuleStates?.[mod.id]?.states[flowModuleStates?.[mod.id]?.job_id]
 		return {
 			type: 'node',
 			id: mod.id,
@@ -334,8 +331,8 @@
 						insertable,
 						insertableEnd,
 						branchable,
-						duration_ms: fms?.duration_ms,
-						bgColor: getStateColor(fms?.type),
+						duration_ms: flowModuleStates?.[mod.id]?.duration_ms,
+						bgColor: getStateColor(flowModuleStates?.[mod.id]?.type),
 						annotation,
 						modules,
 						moving
@@ -376,8 +373,6 @@
 		parent: NestedNodes | string | undefined,
 		loopDepth: number
 	): Loop {
-		let fms = flowModuleStates?.[module.id]?.states[flowModuleStates?.[module.id]?.job_id]
-
 		const loop: Loop = {
 			type: 'loop',
 			items: [
@@ -385,7 +380,9 @@
 					getParentIds(parent),
 					module,
 					undefined,
-					fms?.iteration_total ? 'Iteration ' + fms?.iteration_total : '',
+					flowModuleStates?.[module.id]?.iteration_total
+						? 'Iteration ' + flowModuleStates?.[module.id]?.iteration_total
+						: '',
 					loopDepth,
 					false,
 					false,
@@ -725,8 +722,6 @@
 		const nId = (-idGenerator.next().value - 1 + 1100).toString()
 		parent_module && (errorHandlers[parent_module] = nId)
 		let label = 'Error handler'
-		let fms = flowModuleStates?.[mod.id]?.states[flowModuleStates?.[mod.id]?.job_id]
-
 		return {
 			type: 'node',
 			id: nId,
@@ -738,7 +733,7 @@
 						label,
 						insertable: false,
 						modules: undefined,
-						bgColor: getStateColor(fms?.type),
+						bgColor: getStateColor(flowModuleStates?.[mod.id]?.type),
 						selected: $selectedId == mod.id,
 						index: 0,
 						selectable: true,
